refactor: move code with traits to `src/traits.rs`; move code with directives to `src/directivess`
