+++ src/policy/core.rs
@@ -3,6 +3,7 @@ use html5ever::LocalName;
 use tendril::StrTendril;
 
 use super::builder::PolicyBuilder;
+use crate::traits::{SanitizeDirective, SanitizePolicy};
 use crate::{Permissive, Restrictive};
 
 fn is_node_name_in(names: &[LocalName], node: &NodeRef) -> bool {
@@ -10,95 +11,6 @@ fn is_node_name_in(names: &[LocalName], node: &NodeRef) -> bool {
         return false;
     };
     names.contains(&qual_name.local)
-}
-
-/// A trait for sanitization directives, defines methods for node and attribute sanitization.
-pub trait SanitizeDirective {
-    /// Sanitizes a node by removing elements and attributes based on the policy.
-    fn sanitize_node(policy: &impl SanitizePolicy, node: &NodeRef)
-    where
-        Self: Sized;
-    /// Sanitizes the attributes of a node by removing or retaining them based on the policy.
-    fn sanitize_node_attrs(policy: &impl SanitizePolicy, node: &dom_query::NodeRef)
-    where
-        Self: Sized;
-}
-
-impl SanitizeDirective for Permissive {
-    /// Removes matching elements from the DOM keeping their children.
-    /// Removes matching attributes from the element node.
-    fn sanitize_node(policy: &impl SanitizePolicy, node: &NodeRef) {
-        if policy.is_empty() {
-            return;
-        }
-        let mut next_node = next_child_or_sibling(node, false);
-
-        while let Some(child) = next_node {
-            if policy.should_remove(&child) {
-                next_node = next_child_or_sibling(&child, true);
-                child.remove_from_parent();
-                continue;
-            }
-
-            next_node = next_child_or_sibling(&child, false);
-            if policy.should_exclude(&child) {
-                if let Some(first_inline) = child.first_child() {
-                    child.insert_siblings_before(&first_inline);
-                };
-                child.remove_from_parent();
-            } else {
-                Self::sanitize_node_attrs(policy, &child);
-            }
-        }
-    }
-
-    /// Removes matching attributes from the element node.
-    fn sanitize_node_attrs(policy: &impl SanitizePolicy, node: &dom_query::NodeRef) {
-        if !policy.has_attrs_to_exclude() {
-            return;
-        }
-        policy.exclude_attrs(node, |node, attrs| node.remove_attrs(attrs));
-    }
-}
-
-impl SanitizeDirective for Restrictive {
-    /// Removes elements from the DOM keeping their children with exception of
-    /// elements listed in policy.
-    /// Removes attributes from the element node with exception of
-    /// attributes listed in policy.
-    fn sanitize_node(policy: &impl SanitizePolicy, node: &NodeRef) {
-        let mut next_node = next_child_or_sibling(node, false);
-        while let Some(child) = next_node {
-            
-            if policy.should_remove(&child) {
-                next_node = next_child_or_sibling(&child, true);
-                child.remove_from_parent();
-                continue;
-            }
-            
-            next_node = next_child_or_sibling(&child, false);
-
-            if Self::should_skip(&child) || policy.should_exclude(&child) {
-                Self::sanitize_node_attrs(policy, &child);
-                continue;
-            }
-
-            if let Some(first_inline) = child.first_child() {
-                child.insert_siblings_before(&first_inline);
-            };
-            child.remove_from_parent();
-        }
-    }
-
-    /// Removes all attributes from the element node with exception of
-    /// attributes listed in policy.
-    fn sanitize_node_attrs(policy: &impl SanitizePolicy, node: &dom_query::NodeRef) {
-        if !policy.has_attrs_to_exclude() {
-            node.remove_all_attrs();
-            return;
-        }
-        policy.exclude_attrs(node, |node, attrs| node.retain_attrs(attrs));
-    }
 }
 
 /// An **excluding** rule for sanitizing attributes of a specific element.
@@ -158,21 +70,13 @@ impl<T: SanitizeDirective> Policy<'_, T> {
     }
 }
 
-pub trait SanitizePolicy {
-    fn should_exclude(&self, node: &NodeRef) -> bool;
-    fn should_remove(&self, node: &NodeRef) -> bool;
-    fn has_attrs_to_exclude(&self) -> bool;
-    fn exclude_attrs<F>(&self, node: &NodeRef, exclude_fn: F) where  F: FnOnce(&NodeRef, &[&str]);
-    fn is_empty(&self) -> bool { false }
-}
-
-impl <T: SanitizeDirective>SanitizePolicy for  Policy<'_, T> {
+impl<T: SanitizeDirective> SanitizePolicy for Policy<'_, T> {
     fn should_exclude(&self, node: &NodeRef) -> bool {
-        is_node_name_in(&self.elements_to_exclude, &node) 
+        is_node_name_in(&self.elements_to_exclude, node)
     }
 
     fn should_remove(&self, node: &NodeRef) -> bool {
-        is_node_name_in(&self.elements_to_remove, &node) 
+        is_node_name_in(&self.elements_to_remove, node)
     }
 
     fn has_attrs_to_exclude(&self) -> bool {
@@ -194,7 +98,7 @@ impl <T: SanitizeDirective>SanitizePolicy for  Policy<'_, T> {
             let Some(qual_name) = node.qual_name_ref() else {
                 return;
             };
-    
+
             for rule in &self.attrs_to_exclude {
                 let Some(element_name) = &rule.element else {
                     attrs.extend(rule.attributes.iter());
@@ -205,7 +109,7 @@ impl <T: SanitizeDirective>SanitizePolicy for  Policy<'_, T> {
                 }
             }
         }
-        
+
         exclude_fn(node, &attrs)
     }
 }
@@ -232,23 +136,3 @@ pub type RestrictivePolicy<'a> = Policy<'a, Restrictive>;
 /// Alias for [`RestrictivePolicy`] — denies all elements and attributes by default.
 pub type DenyAllPolicy<'a> = Policy<'a, Restrictive>;
 
-fn next_child_or_sibling<'a>(node: &NodeRef<'a>, ignore_child: bool) -> Option<NodeRef<'a>> {
-    if !ignore_child {
-        if let Some(first_child) = node.first_element_child() {
-            return Some(first_child);
-        }
-    }
-
-    if let Some(sibling) = node.next_element_sibling() {
-        return Some(sibling);
-    } 
-    let mut parent = node.parent();
-    while let Some(parent_node) = parent {
-        if let Some(next_sibling) = parent_node.next_element_sibling() {
-            return Some(next_sibling);
-        } else {
-            parent = parent_node.parent()
-        }
-    }
-    None
-}
